requirements.txt
Django==5.0.1
Faker==18.13.0
black==23.7.0
django-debug-toolbar==4.1.0
djlint===1.34.0
factory-boy==3.2.1
flake8==6.0.0
isort==5.12.0
pytest-cov==4.1.0
pytest-factoryboy==2.5.1
pytest==7.4.0
pytest-django==4.8.0
requests==2.31.0
django-filter==23.5
pillow==10.2.0
pytest-cov==4.1.0
